+++ release_notes.md
@@ -5,6 +5,7 @@
 -   APP-3766 - [API] Added support for Security Labels on Attributes
 -   APP-3775 - [Playbooks] Updated variable pattern regex to better handle mix variables types
 -   APP-3776 - [CLI] Updated spectool for general improvements and fixes
+-   APP-3805 - [API] Added TI Transforms module used to map a TI object to TC Bach or V3 format
 
 ### 3.0.4